+++ src/config.js
@@ -29,8 +29,8 @@ const config = {
 
   hls: {
     root: env.SHLS_MEDIAROOT || "/tmp/hls",
-    hlsTime: 8,
-    hlsListSize: 3,
+    hlsTime: 7,
+    hlsListSize: 2,
   },
 
   stats: {